moved wrap_esp8266 stuff from networking to target
