+++ libs/network/esp8266/jswrap_esp8266_network.h
@@ -35,18 +35,4 @@ JsVar *jswrap_ESP8266_wifi_getDHCPHostname(JsVar *jsCallback);
 void   jswrap_ESP8266_wifi_setDHCPHostname(JsVar *jsHostname);
 void   jswrap_ESP8266_wifi_getHostByName(JsVar *jsHostname, JsVar *jsCallback);
 
-//===== ESP8266 Library
-void   jswrap_ESP8266_dumpAllSocketData();
-void   jswrap_ESP8266_dumpSocket(JsVar *jsSocketId);
-JsVar *jswrap_ESP8266_getResetInfo();
-JsVar *jswrap_ESP8266_getState();
-void   jswrap_ESP8266_logDebug(JsVar *jsDebug);
-void   jswrap_ESP8266_ping(JsVar *jsIpAddr, JsVar *jsPingCallback);
-void   jswrap_ESP8266_reboot();
-void   jswrap_ESP8266_setCPUFreq(JsVar *jsFreq);
-
-void   jswrap_ESP8266_wifi_pre_init();
-void   jswrap_ESP8266_wifi_init();
-void   jswrap_ESP8266_wifi_reset();
-
 #endif /* LIBS_NETWORK_ESP8266_JSWRAP_ESP8266_NETWORK_H_ */
