+++ targets/esp8266/jswrap_esp8266.c
@@ -33,6 +33,7 @@
 typedef long long int64_t;
 
 #include <jswrap_esp8266.h>
+#include "jswrap_esp8266_network.h"
 #include "jsinteractive.h" // Pull inn the jsiConsolePrint function
 
 #define _BV(bit) (1 << (bit))
@@ -44,6 +45,153 @@ static inline uint32_t _getCycleCount(void) {
   __asm__ __volatile__("rsr %0,ccount":"=a" (ccount));
   return ccount;
 }
+
+// ESP8266.reboot
+
+/*JSON{
+  "type"     : "staticmethod",
+  "class"    : "ESP8266",
+  "name"     : "reboot",
+  "generate" : "jswrap_ESP8266_reboot"
+}
+Perform a hardware reset/reboot of the esp8266.
+*/
+void jswrap_ESP8266_reboot() {
+  os_printf("Espruino resetting the esp8266\n");
+  os_delay_us(1000); // time for os_printf to drain
+  system_restart();
+}
+
+//===== ESP8266.getResetInfo
+
+/**
+ * Retrieve the reset information that is stored when the ESP8266 resets.
+ * The result will be a JS object containing the details.
+ */
+/*JSON{
+  "type"     : "staticmethod",
+  "class"    : "ESP8266",
+  "name"     : "getResetInfo",
+  "generate" : "jswrap_ESP8266_getResetInfo",
+  "return"   : ["JsVar","An object with the reset cause information"],
+  "return_object" : "RstInfo"
+}
+At boot time the esp8266's firmware captures the cause of the reset/reboot.  This function returns this information in an object with the following fields:
+
+* `reason`: "power on", "wdt reset", "exception", "soft wdt", "restart", "deep sleep", or "reset pin"
+* `exccause`: exception cause
+* `epc1`, `epc2`, `epc3`: instruction pointers
+* `excvaddr`: address being accessed
+* `depc`: (?)
+
+*/
+JsVar *jswrap_ESP8266_getResetInfo() {
+  struct rst_info* info = system_get_rst_info();
+  JsVar *restartInfo = jspNewObject(NULL, "RstInfo");
+  extern char *rst_codes[]; // in user_main.c
+
+  jsvObjectSetChildAndUnLock(restartInfo, "exccause", jsvNewFromString(rst_codes[info->exccause]));
+  jsvObjectSetChildAndUnLock(restartInfo, "epc1",     jsvNewFromInteger(info->epc1));
+  jsvObjectSetChildAndUnLock(restartInfo, "epc2",     jsvNewFromInteger(info->epc2));
+  jsvObjectSetChildAndUnLock(restartInfo, "epc3",     jsvNewFromInteger(info->epc3));
+  jsvObjectSetChildAndUnLock(restartInfo, "excvaddr", jsvNewFromInteger(info->excvaddr));
+  jsvObjectSetChildAndUnLock(restartInfo, "depc",     jsvNewFromInteger(info->depc));
+  return restartInfo;
+}
+
+//===== ESP8266.logDebug
+
+/*JSON{
+  "type"     : "staticmethod",
+  "class"    : "ESP8266",
+  "name"     : "logDebug",
+  "generate" : "jswrap_ESP8266_logDebug",
+  "params"   : [
+    ["enable", "JsVar", "Enable or disable the debug logging."]
+  ]
+}
+Enable or disable the logging of debug information.  A value of `true` enables debug logging while a value of `false` disables debug logging.  Debug output is sent to UART1 (gpio2).
+ */
+void jswrap_ESP8266_logDebug(
+    JsVar *jsDebug
+  ) {
+  uint8 enable = (uint8)jsvGetBool(jsDebug);
+  os_printf("ESP8255.logDebug, enable=%d\n", enable);
+  system_set_os_print((uint8)jsvGetBool(jsDebug));
+}
+
+//===== ESP8266.setCPUFreq
+
+/*JSON{
+  "type"     : "staticmethod",
+  "class"    : "ESP8266",
+  "name"     : "setCPUFreq",
+  "generate" : "jswrap_ESP8266_setCPUFreq",
+  "params"   : [
+    ["freq", "JsVar", "Desired frequency - either 80 or 160."]
+  ]
+}
+Set the operating frequency of the ESP8266 processor.
+*/
+void jswrap_ESP8266_setCPUFreq(
+    JsVar *jsFreq //!< Operating frequency of the processor.  Either 80 or 160.
+  ) {
+  if (!jsvIsInt(jsFreq)) {
+    jsExceptionHere(JSET_ERROR, "Invalid frequency.");
+    return;
+  }
+  int newFreq = jsvGetInteger(jsFreq);
+  if (newFreq != 80 && newFreq != 160) {
+    jsExceptionHere(JSET_ERROR, "Invalid frequency value, must be 80 or 160.");
+    return;
+  }
+  system_update_cpu_freq(newFreq);
+}
+
+//===== ESP8266.getState
+
+/*JSON{
+  "type"     : "staticmethod",
+  "class"    : "ESP8266",
+  "name"     : "getState",
+  "generate" : "jswrap_ESP8266_getState",
+  "return"   : ["JsVar", "The state of the ESP8266"]
+}
+Returns an object that contains details about the state of the ESP8266 with the following fields:
+
+* `sdkVersion`   - Version of the SDK.
+* `cpuFrequency` - CPU operating frequency in Mhz.
+* `freeHeap`     - Amount of free heap in bytes.
+* `maxCon`       - Maximum number of concurrent connections.
+* `flashMap`     - Configured flash size&map: '512KB:256/256' .. '4MB:512/512'
+* `flashKB`      - Configured flash size in KB as integer
+* `flashChip`    - Type of flash chip as string with manufacturer & chip, ex: '0xEF 0x4016`
+
+*/
+JsVar *jswrap_ESP8266_getState() {
+  // Create a new variable and populate it with the properties of the ESP8266 that we
+  // wish to return.
+  JsVar *esp8266State = jspNewObject(NULL, "ESP8266State");
+  jsvObjectSetChildAndUnLock(esp8266State, "sdkVersion",   jsvNewFromString(system_get_sdk_version()));
+  jsvObjectSetChildAndUnLock(esp8266State, "cpuFrequency", jsvNewFromInteger(system_get_cpu_freq()));
+  jsvObjectSetChildAndUnLock(esp8266State, "freeHeap",     jsvNewFromInteger(system_get_free_heap_size()));
+  jsvObjectSetChildAndUnLock(esp8266State, "maxCon",       jsvNewFromInteger(espconn_tcp_get_max_con()));
+
+  uint32_t map = system_get_flash_size_map();
+  extern char *flash_maps[]; // in user_main.c
+  extern uint16_t flash_kb[]; // in user_main.c
+  jsvObjectSetChildAndUnLock(esp8266State, "flashMap",   jsvNewFromString(flash_maps[map]));
+  jsvObjectSetChildAndUnLock(esp8266State, "flashKB",    jsvNewFromInteger(flash_kb[map]));
+
+  uint32_t fid = spi_flash_get_id();
+  uint32_t chip = (fid&0xff00)|((fid>>16)&0xff);
+  char buff[16];
+  os_sprintf(buff, "0x%02lx 0x%04lx", fid & 0xff, chip);
+  jsvObjectSetChildAndUnLock(esp8266State, "flashChip",   jsvNewFromString(buff));
+  return esp8266State;
+}
+
+//===== ESP8266.neopixelWrite
 
 /*JSON{
  "type"     : "staticmethod",
